upload check file_exist

+++ admin.php
@@ -26,8 +26,14 @@ if (isset($_FILES['testfile'])) {
         draw_admin_form();
     } else {
 
-        $filename = $extension[0] . rand(0, 1000000) . '.' . $extension[1]; //можно было бы и принудительно приписывать .json
+        $filename = $extension[0] . $extension[1]; //можно было бы и принудительно приписывать .json
         //$_FILES['testfile']['name'] = $filename;
+
+        if (file_exists($uploads_dir . '/' . $filename)) {
+            
+            echo 'Такой файл теста уже существует !!!';
+            die();
+        }
 
         if (move_uploaded_file($_FILES['testfile']['tmp_name'], $uploads_dir . '/' . $filename)) {
